requirements.txt
# FastAPI + async
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pgvector>=0.3.6

# Redis
redis>=5.2.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# OpenAI
openai>=1.55.0

# State encoder (sentence-transformers)
sentence-transformers>=3.3.0

# Validation & config
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
email-validator>=2.0.0

# Observability
structlog>=24.4.0
prometheus-client>=0.21.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
aiosqlite>=0.20.0

# Utilities
msgpack>=1.0.0
